+++ network/index.js
@@ -16,6 +16,7 @@ const reconnect = require('./reconnect');
 const removePeer = require('./remove_peer');
 const sendGift = require('./send_gift');
 const setAutopilot = require('./set_autopilot');
+const transferFunds = require('./transfer_funds');
 
 module.exports = {
   currencyForNetwork,
@@ -36,4 +37,5 @@ module.exports = {
   removePeer,
   sendGift,
   setAutopilot,
+  transferFunds,
 };